Add files via upload
As I was first introduced to Java programming language, I took a first step learning the syntax by writing this game. 
This is called a guessing game ,where a player is given counted turns to guess a correct word and gets terminated when their turns runs out and the name gets revealed.