Add files via upload
HANGMAN is the name of the game. very simple as i was developing it.
It gives a user an incomplete name ( F___) and counted turns to complete it. The user must complete it before it runs out of turns,
if times runs out, the game ends and the correct name is shown.