Add files via upload
I took a course and earned a certificate in Web Development. This project is the results of instructions that were given to me to create a mini-tour guide about places of choice. In the home page I have pictures of the places that are clickable when you hover on them. Then you get into dipper details inside in the page of each city.
The very interesting part is that there is a link at the bottom of the page, that you click and leads you to the full web of the place to get answers about the place.